Views

+++ TeamRoles/Controllers/RolesController.cs
@@ -19,11 +19,11 @@ namespace TeamRoles.Controllers
             _userManager = new UserManager<ApplicationUser>(new UserStore<ApplicationUser>(db));
         }
 
-//        // GET: Roles
-//        public ActionResult Index()
-//        {
-//            return View(db.Roles.ToList());
-//        }
+        //        // GET: Roles
+        //        public ActionResult Index()
+        //        {
+        //            return View(db.Roles.ToList());
+        //        }
 
 
         public ActionResult Create()
@@ -51,27 +51,33 @@ namespace TeamRoles.Controllers
         public ActionResult Delete(string roleName)
         {
             var list = db.Roles.OrderBy(r => r.Name).ToList()
-                .Select(rr => new SelectListItem {Value = rr.Name.ToString(), Text = rr.Name}).ToList();
+                .Select(rr => new SelectListItem { Value = rr.Name.ToString(), Text = rr.Name }).ToList();
             ViewBag.Roles = list;
 
             var thisRole =
                 db.Roles.FirstOrDefault(r => r.Name.Equals(roleName, StringComparison.CurrentCultureIgnoreCase));
+
+            if (thisRole == null || list == null)
+            {
+                return View();
+            }
+
             db.Roles.Remove(thisRole);
             db.SaveChanges();
 
-            ViewBag.DeleteMessage = "Role Deleted";
+            //ViewBag.SuccessMessage = "Success!";
 
             return RedirectToAction("ManageUsers");
         }
 
 
-//        // GET: /Roles/Edit/5
-//        public ActionResult Edit(string roleName)
-//        {
-//            var thisRole = db.Roles.Where(r => r.Name.Equals(roleName, StringComparison.CurrentCultureIgnoreCase)).FirstOrDefault();
-//
-//            return View(thisRole);
-//        }
+        //        // GET: /Roles/Edit/5
+        //        public ActionResult Edit(string roleName)
+        //        {
+        //            var thisRole = db.Roles.Where(r => r.Name.Equals(roleName, StringComparison.CurrentCultureIgnoreCase)).FirstOrDefault();
+        //
+        //            return View(thisRole);
+        //        }
 
         /*//
         // POST: /Roles/Edit/5
@@ -97,7 +103,7 @@ namespace TeamRoles.Controllers
         {
             // prepopulat roles for the view dropdown
             var list = db.Roles.OrderBy(r => r.Name).ToList().Select(rr =>
-                new SelectListItem {Value = rr.Name.ToString(), Text = rr.Name}).ToList();
+                new SelectListItem { Value = rr.Name.ToString(), Text = rr.Name }).ToList();
             ViewBag.Roles = list;
             return View();
         }
@@ -110,16 +116,22 @@ namespace TeamRoles.Controllers
                 u.UserName.Equals(UserName, StringComparison.CurrentCultureIgnoreCase));
 
             var list = db.Roles.OrderBy(r => r.Name).ToList()
-                .Select(rr => new SelectListItem {Value = rr.Name.ToString(), Text = rr.Name}).ToList();
+                .Select(rr => new SelectListItem { Value = rr.Name.ToString(), Text = rr.Name }).ToList();
             ViewBag.Roles = list;
 
-//            if (user == null || RoleName == null)
-//            {
-//                ViewBag.Message = "Role cannot created !";
-//                return View("ManageUsers");
-//            }
-
             var manager = new ApplicationUserManager(new UserStore<ApplicationUser>(db));
+
+            if (user == null || RoleName == null)
+            {
+                ViewBag.Message = "Role cannot created !";
+                return View("ManageUsers");
+            }
+            else if (manager.IsInRole(user.Id, RoleName))
+            {
+                ViewBag.RoleMessage = "User has already this role";
+                return View("ManageUsers");
+            }
+
             manager.AddToRole(user.Id, RoleName);
 
             return View("ManageUsers");
@@ -129,21 +141,24 @@ namespace TeamRoles.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult GetRoles(string UserName)
         {
-            if (!string.IsNullOrWhiteSpace(UserName)) //TODO Advise
+            var user = db.Users.FirstOrDefault(u => u.UserName.Equals(UserName, StringComparison.CurrentCultureIgnoreCase));
+
+            var list = db.Roles.OrderBy(r => r.Name)
+                    .ToList()
+                    .Select(rr => new SelectListItem { Value = rr.Name.ToString(), Text = rr.Name })
+                    .ToList();
+            ViewBag.Roles = list;
+
+            if (user == null || list == null)
             {
-                var user = db.Users.FirstOrDefault(u =>
-                    u.UserName.Equals(UserName, StringComparison.CurrentCultureIgnoreCase));
-                var manager = new ApplicationUserManager(new UserStore<ApplicationUser>(db));
-
-                ViewBag.RolesForThisUser = manager.GetRoles(user.Id);
-
-                // prepopulat roles for the view dropdown
-                var list = db.Roles.OrderBy(r => r.Name).ToList()
-                    .Select(rr => new SelectListItem {Value = rr.Name.ToString(), Text = rr.Name}).ToList();
-                ViewBag.Roles = list;
+                ViewBag.ResultMessage = "Cannot view Roles !";
             }
 
-//            ViewBag.ResultMessage = "Cannot view Roles !";
+            else if (!string.IsNullOrWhiteSpace(UserName))
+            {
+                var manager = new ApplicationUserManager(new UserStore<ApplicationUser>(db));
+                ViewBag.RolesForThisUser = manager.GetRoles(user.Id);
+            }
 
             return View("ManageUsers");
         }
@@ -158,7 +173,7 @@ namespace TeamRoles.Controllers
             var user = db.Users.FirstOrDefault(u =>
                 u.UserName.Equals(UserName, StringComparison.CurrentCultureIgnoreCase));
 
-            /*if (user == null || RoleName == null)
+            if (user == null || RoleName == null)
             {
                 ViewBag.Message1 = "Wrong check again !";
             }
@@ -170,10 +185,10 @@ namespace TeamRoles.Controllers
             else
             {
                 ViewBag.Message3 = "This user doesn't belong to selected role.";
-            }*/
+            }
             // prepopulat roles for the view dropdown
             var list = db.Roles.OrderBy(r => r.Name).ToList()
-                .Select(rr => new SelectListItem {Value = rr.Name.ToString(), Text = rr.Name}).ToList();
+                .Select(rr => new SelectListItem { Value = rr.Name.ToString(), Text = rr.Name }).ToList();
             ViewBag.Roles = list;
 
             return View("ManageUsers");
